Moved button to lib
