+++ frontend/src/lib/Button.js
@@ -1,4 +1,4 @@
-gitimport React from 'react';
+import React from 'react';
 import styled from 'styled-components';
 
 const PurpleButton = styled.button`
